finally sorted out the fbx import
previous problems were due to a combination of: I forgot that almost every 3d export format exports rotations in radians, and also I was doing my transformation multiplication of rotation matrices in the wrong order. need to check to see if fbx always wants zyx order, or if there is a flag I need to read to detect which order.