obj loading calculates tangents
calculating tangents and bitangents for objs, seeeems to be working? For a little bit this had some mysterious flipping of the y direction of the light when using these for normal map reading, and am frustrated that I don't know what change fixed it. should try and track it down so I understand better.