fixed the bug in the tangents & normals
turns out the bug was fixing itself any time the objects were loading from .pmesh instead of streamed direct from .obj. this has something to do with the fact that I was creating the mesh class before assigning tangent data to its vertices. by changing the order so that I calculate the tangent data along with parsing the vertex data out of the obj file, and assigning it all to vertices that are used to instantiate the mesh, I seem to have fixed the bug. I think this has something to do with how opengl was reserving the buffers, which currently occurs when a mesh object is created, but I'm not certain.