still tinkerning with normal mapping
moved more of the tangent math out to asset loader, calculating the face tangent only once then the vertex orthagonalizes it with its normal and creates the bitangents. most of it seems to work except one face on my cube.obj. it renders correctly in blender so the problem is definitely in the pixie engine somewhere